+++ fireREST/__init__.py
@@ -1,20 +1,32 @@
+# -*- coding: utf-8 -*-
+
 import json
 import logging
 import requests
 import urllib3
 
-from .mapping import OBJECT_TYPE
+from .decorators import (
+    cache_result,
+    log_request,
+    minimum_version_required,
+    validate_object_type,
+)
+from .exceptions import (
+    GenericApiError,
+    InvalidNamespaceError,
+    AuthError,
+    AuthRefreshError,
+    RateLimitException,
+)
 from .version import __version__
 
 from copy import deepcopy
-from functools import lru_cache, wraps
-from http.client import responses as http_responses
 from packaging import version
 from requests.auth import HTTPBasicAuth
 from requests.exceptions import ConnectionError
 from time import sleep
 from typing import Dict, List
-from uuid import uuid4, UUID
+from uuid import UUID
 
 #: protocol used to connect to api
 API_PROTOCOL = 'https'
@@ -51,160 +63,6 @@ API_DEFAULT_DOMAIN = 'Global'
 
 #: retry timer in seconds for repeating failed api calls
 API_RETRY_TIMER = 10
-
-
-class GenericApiError(Exception):
-    '''
-    generic api error exception
-    '''
-
-    def __init__(self, message):
-        super().__init__(message)
-
-
-class InvalidNamespaceError(Exception):
-    '''
-    exeption thrown when invalid namespace is being passed to api
-    '''
-
-    def __init__(self, message):
-        super().__init__(message)
-
-
-class AuthError(Exception):
-    '''
-    generic api authentication failure exception
-    '''
-
-    def __init__(self, message):
-        super().__init__(message)
-
-
-class AuthRefreshError(Exception):
-    '''
-    exception used when api token refresh fails
-    '''
-
-    def __init__(self, message):
-        super().__init__(message)
-
-
-class RateLimitException(Exception):
-    '''
-    exception used when fmc rate limiter kicks in
-    '''
-
-    def __init__(self, message):
-        super().__init__(message)
-
-
-class UnsupportedOperationError(Exception):
-    '''
-    exception used when unsupported operation is being performed
-    '''
-
-    def __init__(self, message):
-        super().__init__(message)
-
-
-class UnsupportedObjectTypeError(Exception):
-    '''
-    exception used when unsupported object type is being used
-    '''
-
-    def __init__(self, message):
-        super().__init__(message)
-
-
-class LogRequest(object):
-    '''
-    decorator that adds additional debug logging for api requests
-    '''
-
-    def __init__(self, action):
-        self.action = action
-
-    def __call__(self, f):
-        def wrapped_f(*args):
-            action = self.action
-            logger = args[0].logger
-            request = args[1]
-            request_id = str(uuid4())[:8]
-            try:
-                data = args[3]
-            except IndexError:
-                data = None
-            logger.debug(f'[{request_id}] [{action}] {request}')
-            if data:
-                logger.debug(f'[{request_id}] [Data] {data}')
-            result = f(*args)
-            status_code = result.status_code
-            status_code_name = http_responses[status_code]
-            logger.debug(f'[{request_id}] [Response] {status_code_name} ({status_code})')
-            if status_code >= 299:
-                logger.debug(f'[{request_id}] [Message] {result.content}')
-            return result
-        return wrapped_f
-
-
-class MinimumVersionRequired(object):
-    '''
-    decorator that specifies the minimal required software version to use the operation
-    '''
-
-    def __init__(self, minimum_version):
-        self.minimum_version = version.parse(minimum_version)
-
-    def __call__(self, f):
-        def wrapped_f(*args, **kwargs):
-            minimum_version = self.minimum_version
-            try:
-                installed_version = args[0].version
-            except AttributeError:
-                return f(*args, **kwargs)
-            if installed_version < minimum_version:
-                raise UnsupportedOperationError(
-                    f'{f.__name__} requires fmc software version {minimum_version}. Installed version: {installed_version}')
-            return f(*args, **kwargs)
-        return wrapped_f
-
-
-def CacheResult(f):
-    '''
-    decorator that applies functools lru_cache if cache is enabled in Client object
-    '''
-    @wraps(f)
-    def wrapper(*args, **kwargs):
-        enabled = args[0].cache
-        if enabled:
-            @lru_cache(maxsize=256)
-            def cached_wrapper():
-                return f(*args, **kwargs)
-            return cached_wrapper
-        return f(*args, **kwargs)
-    return wrapper
-
-
-def ValidateObjectType(f):
-    '''
-    decorator that validates object type and transforms input if neccessary
-    '''
-    @wraps(f)
-    def wrapper(*args, **kwargs):
-        try:
-            object_type = kwargs.get('object_type', None)
-            if object_type:
-                kwargs['object_type'] = OBJECT_TYPE[object_type]
-            else:
-                args = list(args)
-                object_type = args[1]
-                args[1] = OBJECT_TYPE[object_type]
-        except KeyError:
-            if object_type in OBJECT_TYPE.values():
-                return f(*args, **kwargs)
-            raise UnsupportedObjectTypeError(f'{object_type} is not a valid object type')
-        return f(*args, **kwargs)
-    return wrapper
 
 
 class Client(object):
@@ -407,7 +265,7 @@ class Client(object):
 
         self.logger.debug(f'Successfully refreshed authorization token for {self.hostname}')
 
-    @LogRequest('DELETE')
+    @log_request('DELETE')
     def _delete(self, request: str, params=None):
         '''
         DELETE operation
@@ -432,7 +290,7 @@ class Client(object):
             return self._delete(request, params)
         return response
 
-    @LogRequest('GET')
+    @log_request('GET')
     def _get_request(self, request: str, params=None):
         '''
         GET operation without paging support
@@ -481,7 +339,7 @@ class Client(object):
                 responses.append(response_page)
         return self._squash_responses(responses)
 
-    @LogRequest('POST')
+    @log_request('POST')
     def _post(self, request: str, data: Dict, params=None):
         '''
         POST operation
@@ -508,7 +366,7 @@ class Client(object):
             return self._post(request, data, params)
         return response
 
-    @LogRequest('PUT')
+    @log_request('PUT')
     def _put(self, request: str, data: Dict, params=None):
         '''
         PUT operation
@@ -549,7 +407,7 @@ class Client(object):
         sanitized_payload.pop('id', None)
         return sanitized_payload
 
-    @MinimumVersionRequired('6.1.0')
+    @minimum_version_required('6.1.0')
     def get_object_id_by_name(self, object_type: str, object_name: str):
         '''
         helper function to retrieve object id by name
@@ -565,7 +423,7 @@ class Client(object):
                 return obj['id']
         return None
 
-    @MinimumVersionRequired('6.1.0')
+    @minimum_version_required('6.1.0')
     def get_device_id_by_name(self, device_name: str):
         '''
         helper function to retrieve device id by name
@@ -580,7 +438,7 @@ class Client(object):
                 return device['id']
         return None
 
-    @MinimumVersionRequired('6.2.3')
+    @minimum_version_required('6.2.3')
     def get_device_hapair_id_by_name(self, device_hapair_name: str):
         '''
         helper function to retrieve device ha - pair id by name
@@ -595,7 +453,7 @@ class Client(object):
                 return ha_pair['id']
         return None
 
-    @MinimumVersionRequired('6.2.3')
+    @minimum_version_required('6.2.3')
     def get_device_id_from_hapair(self, device_hapair_id: str):
         '''
         helper function to retrieve device id from ha - pair
@@ -607,7 +465,7 @@ class Client(object):
         ha_pair = self._get(url)
         return ha_pair['primary']['id']
 
-    @MinimumVersionRequired('6.2.3')
+    @minimum_version_required('6.2.3')
     def get_nat_policy_id_by_name(self, nat_policy_name: str):
         '''
         helper function to retrieve nat policy id by name
@@ -622,7 +480,7 @@ class Client(object):
                 return nat_policy['id']
         return None
 
-    @MinimumVersionRequired('6.1.0')
+    @minimum_version_required('6.1.0')
     def get_acp_id_by_name(self, policy_name: str):
         '''
         helper function to retrieve access control policy id by name
@@ -637,7 +495,7 @@ class Client(object):
                 return accesspolicy['id']
         return None
 
-    @MinimumVersionRequired('6.1.0')
+    @minimum_version_required('6.1.0')
     def get_acp_rule_id_by_name(self, policy_name: str, rule_name: str):
         '''
         helper function to retrieve access control policy rule id by name
@@ -654,7 +512,7 @@ class Client(object):
                 return acp_rule['id']
         return None
 
-    @MinimumVersionRequired('6.1.0')
+    @minimum_version_required('6.1.0')
     def get_syslog_alert_id_by_name(self, syslog_alert_name: str):
         '''
         helper function to retrieve syslog alert object id by name
@@ -667,7 +525,7 @@ class Client(object):
                 return syslog_alert['id']
         return None
 
-    @MinimumVersionRequired('6.1.0')
+    @minimum_version_required('6.1.0')
     def get_snmp_alert_id_by_name(self, snmp_alert_name: str):
         '''
         helper function to retrieve snmp alert object id by name
@@ -708,39 +566,39 @@ class Client(object):
         self.logger.debug(f'Available Domains: {available_domains}')
         return None
 
-    @MinimumVersionRequired('6.1.0')
+    @minimum_version_required('6.1.0')
     def get_system_version(self):
         request = '/info/serverversion'
         url = self._url('platform', request)
         return self._get(url)
 
-    @MinimumVersionRequired('6.1.0')
+    @minimum_version_required('6.1.0')
     def get_audit_records(self):
         request = '/audit/auditrecords'
         url = self._url('platform', request)
         return self._get(url)
 
-    @MinimumVersionRequired('6.1.0')
+    @minimum_version_required('6.1.0')
     def get_syslogalerts(self):
         request = '/policy/syslogalerts'
         url = self._url('config', request)
         return self._get(url)
 
-    @MinimumVersionRequired('6.1.0')
+    @minimum_version_required('6.1.0')
     def get_snmpalerts(self):
         request = '/policy/snmpalerts'
         url = self._url('config', request)
         return self._get(url)
 
-    @MinimumVersionRequired('6.1.0')
-    @ValidateObjectType
+    @minimum_version_required('6.1.0')
+    @validate_object_type
     def create_object(self, object_type: str, data: Dict):
         request = f'/object/{object_type}'
         url = self._url('config', request)
         return self._post(url, data)
 
-    @MinimumVersionRequired('6.1.0')
-    @ValidateObjectType
+    @minimum_version_required('6.1.0')
+    @validate_object_type
     def get_objects(self, object_type: str):
         request = f'/object/{object_type}'
         url = self._url('config', request)
@@ -749,8 +607,8 @@ class Client(object):
         }
         return self._get(url, params)
 
-    @MinimumVersionRequired('6.4.0')
-    @ValidateObjectType
+    @minimum_version_required('6.4.0')
+    @validate_object_type
     def get_objects_override(self, object_type: str, objects: List):
         overrides = list()
         for obj in objects:
@@ -759,15 +617,15 @@ class Client(object):
                 overrides.extend(responses)
         return overrides
 
-    @MinimumVersionRequired('6.1.0')
-    @ValidateObjectType
+    @minimum_version_required('6.1.0')
+    @validate_object_type
     def get_object(self, object_type: str, object_id: str):
         request = f'/object/{object_type}/{object_id}'
         url = self._url('config', request)
         return self._get(url)
 
-    @MinimumVersionRequired('6.4.0')
-    @ValidateObjectType
+    @minimum_version_required('6.4.0')
+    @validate_object_type
     def get_object_override(self, object_type: str, object_id: str):
         request = f'/object/{object_type}/{object_id}/overrides'
         url = self._url('config', request)
@@ -776,27 +634,27 @@ class Client(object):
         }
         return self._get(url, params)
 
-    @MinimumVersionRequired('6.1.0')
-    @ValidateObjectType
+    @minimum_version_required('6.1.0')
+    @validate_object_type
     def update_object(self, object_type: str, object_id: str, data: Dict):
         request = f'/object/{object_type}/{object_id}'
         url = self._url('config', request)
         return self._put(url, data)
 
-    @MinimumVersionRequired('6.1.0')
-    @ValidateObjectType
+    @minimum_version_required('6.1.0')
+    @validate_object_type
     def delete_object(self, object_type: str, object_id: str):
         request = f'/object/{object_type}/{object_id}'
         url = self._url('config', request)
         return self._delete(url)
 
-    @MinimumVersionRequired('6.1.0')
+    @minimum_version_required('6.1.0')
     def create_device(self, data: Dict):
         request = '/devices/devicerecords'
         url = self._url('config', request)
         return self._post(url, data)
 
-    @MinimumVersionRequired('6.1.0')
+    @minimum_version_required('6.1.0')
     def get_devices(self):
         request = '/devices/devicerecords'
         url = self._url('config', request)
@@ -805,25 +663,25 @@ class Client(object):
         }
         return self._get(url, params)
 
-    @MinimumVersionRequired('6.1.0')
+    @minimum_version_required('6.1.0')
     def get_device(self, device_id: str):
         request = f'/devices/devicerecords/{device_id}'
         url = self._url('config', request)
         return self._get(url)
 
-    @MinimumVersionRequired('6.1.0')
+    @minimum_version_required('6.1.0')
     def update_device(self, device_id: str, data: Dict):
         request = f'/devices/devicerecords/{device_id}'
         url = self._url('config', request)
         return self._put(url, data)
 
-    @MinimumVersionRequired('6.1.0')
+    @minimum_version_required('6.1.0')
     def delete_device(self, device_id: str):
         request = f'/devices/devicerecords/{device_id}'
         url = self._url('config', request)
         return self._delete(url)
 
-    @MinimumVersionRequired('6.2.3')
+    @minimum_version_required('6.2.3')
     def get_device_hapairs(self):
         request = '/devicehapairs/ftddevicehapairs'
         params = {
@@ -832,31 +690,31 @@ class Client(object):
         url = self._url('config', request)
         return self._get(url, params)
 
-    @MinimumVersionRequired('6.2.3')
+    @minimum_version_required('6.2.3')
     def create_device_hapair(self, data: Dict):
         request = '/devicehapairs/ftddevicehapairs/{}'
         url = self._url('config', request)
         return self._get(url, data)
 
-    @MinimumVersionRequired('6.2.3')
+    @minimum_version_required('6.2.3')
     def get_device_hapair(self, device_hapair_id: str):
         request = f'/devicehapairs/ftddevicehapairs/{device_hapair_id}'
         url = self._url('config', request)
         return self._get(url)
 
-    @MinimumVersionRequired('6.2.3')
+    @minimum_version_required('6.2.3')
     def update_device_hapair(self, data: Dict, device_hapair_id: str):
         request = f'/devicehapairs/ftddevicehapairs/{device_hapair_id}'
         url = self._url('config', request)
         return self._put(url, data)
 
-    @MinimumVersionRequired('6.2.3')
+    @minimum_version_required('6.2.3')
     def delete_device_hapair(self, device_hapair_id: str):
         request = f'/devicehapairs/ftddevicehapairs/{device_hapair_id}'
         url = self._url('config', request)
         return self._delete(url)
 
-    @MinimumVersionRequired('6.3.0')
+    @minimum_version_required('6.3.0')
     def get_device_hapair_monitoredinterfaces(self, device_hapair_id: str):
         request = f'/devicehapairs/ftddevicehapairs/{device_hapair_id}/monitoredinterfaces'
         params = {
@@ -865,19 +723,19 @@ class Client(object):
         url = self._url('config', request)
         return self._get(url, params)
 
-    @MinimumVersionRequired('6.3.0')
+    @minimum_version_required('6.3.0')
     def get_device_hapair_monitoredinterface(self, device_hapair_id: str, monitoredinterface_id: str):
         request = f'/devicehapairs/ftddevicehapairs/{device_hapair_id}/monitoredinterfaces/{monitoredinterface_id}'
         url = self._url('config', request)
         return self._get(url)
 
-    @MinimumVersionRequired('6.3.0')
+    @minimum_version_required('6.3.0')
     def update_device_hapair_monitoredinterface(self, device_hapair_id: str, monitoredinterface_id: str, data: Dict):
         request = f'/devicehapairs/ftddevicehapairs/{device_hapair_id}/monitoredinterfaces/{monitoredinterface_id}'
         url = self._url('config', request)
         return self._put(url, data)
 
-    @MinimumVersionRequired('6.1.0')
+    @minimum_version_required('6.1.0')
     def get_ftd_physical_interfaces(self, device_id: str):
         request = f'/devices/devicerecords/{device_id}/physicalinterfaces'
         url = self._url('config', request)
@@ -886,25 +744,25 @@ class Client(object):
         }
         return self._get(url, params)
 
-    @MinimumVersionRequired('6.1.0')
+    @minimum_version_required('6.1.0')
     def get_ftd_physical_interface(self, device_id: str, interface_id: str):
         request = f'/devices/devicerecords/{device_id}/physicalinterfaces/{interface_id}'
         url = self._url('config', request)
         return self._get(url)
 
-    @MinimumVersionRequired('6.1.0')
+    @minimum_version_required('6.1.0')
     def update_ftd_physical_interface(self, device_id: str, data: Dict):
         request = f'/devices/devicerecords/{device_id}/physicalinterfaces'
         url = self._url('config', request)
         return self._put(url, data)
 
-    @MinimumVersionRequired('6.1.0')
+    @minimum_version_required('6.1.0')
     def create_ftd_redundant_interface(self, device_id: str, data: Dict):
         request = f'/devices/devicerecords/{device_id}/redundantinterfaces'
         url = self._url('config', request)
         return self._post(url, data)
 
-    @MinimumVersionRequired('6.1.0')
+    @minimum_version_required('6.1.0')
     def get_ftd_redundant_interfaces(self, device_id: str):
         request = f'/devices/devicerecords/{device_id}/redundantinterfaces'
         url = self._url('config', request)
@@ -913,31 +771,31 @@ class Client(object):
         }
         return self._get(url, params)
 
-    @MinimumVersionRequired('6.1.0')
+    @minimum_version_required('6.1.0')
     def get_ftd_redundant_interface(self, device_id: str, interface_id: str):
         request = f'/devices/devicerecords/{device_id}/redundantinterfaces/{interface_id}'
         url = self._url('config', request)
         return self._get(url)
 
-    @MinimumVersionRequired('6.1.0')
+    @minimum_version_required('6.1.0')
     def update_ftd_redundant_interface(self, device_id: str, data: Dict):
         request = f'/devices/devicerecords/{device_id}/redundantinterfaces'
         url = self._url('config', request)
         return self._put(url, data)
 
-    @MinimumVersionRequired('6.1.0')
+    @minimum_version_required('6.1.0')
     def delete_ftd_redundant_interface(self, device_id: str, interface_id: str):
         request = f'/devices/devicerecords/{device_id}/redundantinterfaces/{interface_id}'
         url = self._url('config', request)
         return self._delete(url)
 
-    @MinimumVersionRequired('6.1.0')
+    @minimum_version_required('6.1.0')
     def create_ftd_portchannel_interface(self, device_id: str, data: Dict):
         request = f'/devices/devicerecords/{device_id}/etherchannelinterfaces'
         url = self._url('config', request)
         return self._post(url, data)
 
-    @MinimumVersionRequired('6.1.0')
+    @minimum_version_required('6.1.0')
     def get_ftd_portchannel_interfaces(self, device_id: str):
         request = f'/devices/devicerecords/{device_id}/etherchannelinterfaces'
         url = self._url('config', request)
@@ -946,31 +804,31 @@ class Client(object):
         }
         return self._get(url, params)
 
-    @MinimumVersionRequired('6.1.0')
+    @minimum_version_required('6.1.0')
     def get_ftd_portchannel_interface(self, device_id: str, interface_id: str):
         request = f'/devices/devicerecords/{device_id}/etherchannelinterfaces/{interface_id}'
         url = self._url('config', request)
         return self._get(url)
 
-    @MinimumVersionRequired('6.1.0')
+    @minimum_version_required('6.1.0')
     def update_ftd_portchannel_interface(self, device_id: str, data: Dict):
         request = f'/devices/devicerecords/{device_id}/etherchannelinterfaces'
         url = self._url('config', request)
         return self._put(url, data)
 
-    @MinimumVersionRequired('6.1.0')
+    @minimum_version_required('6.1.0')
     def delete_ftd_portchannel_interface(self, device_id: str, interface_id: str):
         request = f'/devices/devicerecords/{device_id}/etherchannelinterfaces/{interface_id}'
         url = self._url('config', request)
         return self._delete(url)
 
-    @MinimumVersionRequired('6.1.0')
+    @minimum_version_required('6.1.0')
     def create_ftd_sub_interface(self, device_id: str, data: Dict):
         request = f'/devices/devicerecords/{device_id}/subinterfaces'
         url = self._url('config', request)
         return self._post(url, data)
 
-    @MinimumVersionRequired('6.1.0')
+    @minimum_version_required('6.1.0')
     def get_ftd_sub_interfaces(self, device_id: str):
         request = f'/devices/devicerecords/{device_id}/subinterfaces'
         url = self._url('config', request)
@@ -979,31 +837,31 @@ class Client(object):
         }
         return self._get(url, params)
 
-    @MinimumVersionRequired('6.1.0')
+    @minimum_version_required('6.1.0')
     def get_ftd_sub_interface(self, device_id: str, interface_id: str):
         request = f'/devices/devicerecords/{device_id}/subinterfaces/{interface_id}'
         url = self._url('config', request)
         return self._get(url)
 
-    @MinimumVersionRequired('6.1.0')
+    @minimum_version_required('6.1.0')
     def update_ftd_sub_interface(self, device_id: str, interface_id: str, data: Dict):
         request = f'/devices/devicerecords/{device_id}/subinterfaces/{interface_id}'
         url = self._url('config', request)
         return self._put(url, data)
 
-    @MinimumVersionRequired('6.1.0')
+    @minimum_version_required('6.1.0')
     def delete_ftd_sub_interface(self, device_id: str, interface_id: str):
         request = f'/devices/devicerecords/{device_id}/subinterfaces/{interface_id}'
         url = self._url('config', request)
         return self._delete(url)
 
-    @MinimumVersionRequired('6.3.0')
+    @minimum_version_required('6.3.0')
     def create_ftd_ipv4staticroute(self, device_id: str, data: Dict):
         request = f'/devices/devicerecords/{device_id}/routing/ipv4staticroutes'
         url = self._url('config', request)
         return self._post(url, data)
 
-    @MinimumVersionRequired('6.3.0')
+    @minimum_version_required('6.3.0')
     def get_ftd_ipv4staticroutes(self, device_id: str):
         request = f'/devices/devicerecords/{device_id}/routing/ipv4staticroutes'
         url = self._url('config', request)
@@ -1012,31 +870,31 @@ class Client(object):
         }
         return self._get(url, params)
 
-    @MinimumVersionRequired('6.3.0')
+    @minimum_version_required('6.3.0')
     def get_ftd_ipv4staticroute(self, device_id: str, route_id: str):
         request = f'/devices/devicerecords/{device_id}/routing/ipv4staticroutes/{route_id}'
         url = self._url('config', request)
         return self._get(url)
 
-    @MinimumVersionRequired('6.3.0')
+    @minimum_version_required('6.3.0')
     def update_ftd_ipv4staticroute(self, device_id: str, route_id: str, data: Dict):
         request = f'/devices/devicerecords/{device_id}/routing/ipv4staticroutes/{route_id}'
         url = self._url('config', request)
         return self._put(url, data)
 
-    @MinimumVersionRequired('6.3.0')
+    @minimum_version_required('6.3.0')
     def delete_ftd_ipv4staticroute(self, device_id: str, route_id: str):
         request = f'/devices/devicerecords/{device_id}/routing/ipv4staticroutes/{route_id}'
         url = self._url('config', request)
         return self._delete(url)
 
-    @MinimumVersionRequired('6.3.0')
+    @minimum_version_required('6.3.0')
     def create_ftd_ipv6staticroute(self, device_id: str, data: Dict):
         request = f'/devices/devicerecords/{device_id}/routing/ipv6staticroutes'
         url = self._url('config', request)
         return self._post(url, data)
 
-    @MinimumVersionRequired('6.3.0')
+    @minimum_version_required('6.3.0')
     def get_ftd_ipv6staticroutes(self, device_id: str):
         request = f'/devices/devicerecords/{device_id}/routing/ipv6staticroutes'
         url = self._url('config', request)
@@ -1045,31 +903,31 @@ class Client(object):
         }
         return self._get(url, params)
 
-    @MinimumVersionRequired('6.3.0')
+    @minimum_version_required('6.3.0')
     def get_ftd_ipv6staticroute(self, device_id: str, route_id: str):
         request = f'/devices/devicerecords/{device_id}/routing/ipv6staticroutes/{route_id}'
         url = self._url('config', request)
         return self._get(url)
 
-    @MinimumVersionRequired('6.3.0')
+    @minimum_version_required('6.3.0')
     def update_ftd_ipv6staticroute(self, device_id: str, route_id: str, data: Dict):
         request = f'/devices/devicerecords/{device_id}/routing/ipv6staticroutes/{route_id}'
         url = self._url('config', request)
         return self._put(url, data)
 
-    @MinimumVersionRequired('6.3.0')
+    @minimum_version_required('6.3.0')
     def delete_ftd_ipv6staticroute(self, device_id: str, route_id: str):
         request = f'/devices/devicerecords/{device_id}/routing/ipv6staticroutes/{route_id}'
         url = self._url('config', request)
         return self._delete(url)
 
-    @MinimumVersionRequired('6.1.0')
+    @minimum_version_required('6.1.0')
     def deploy(self, data: Dict):
         request = '/deployment/deploymentrequests'
         url = self._url('config', request)
         return self._post(url, data)
 
-    @MinimumVersionRequired('6.1.0')
+    @minimum_version_required('6.1.0')
     def get_deployable_devices(self):
         request = '/deployment/deployabledevices'
         url = self._url('config', request)
@@ -1078,13 +936,13 @@ class Client(object):
         }
         return self._get(url, params)
 
-    @MinimumVersionRequired('6.1.0')
+    @minimum_version_required('6.1.0')
     def create_policy(self, policy_type: str, data: Dict):
         request = f'/policy/{policy_type}'
         url = self._url('config', request)
         return self._post(url, data)
 
-    @MinimumVersionRequired('6.1.0')
+    @minimum_version_required('6.1.0')
     def get_policies(self, policy_type: str):
         request = f'/policy/{policy_type}'
         url = self._url('config', request)
@@ -1093,7 +951,7 @@ class Client(object):
         }
         return self._get(url, params)
 
-    @MinimumVersionRequired('6.1.0')
+    @minimum_version_required('6.1.0')
     def get_policy(self, policy_id: str, policy_type: str):
         request = f'/policy/{policy_type}/{policy_id}'
         params = {
@@ -1102,19 +960,19 @@ class Client(object):
         url = self._url('config', request)
         return self._get(url, params)
 
-    @MinimumVersionRequired('6.1.0')
+    @minimum_version_required('6.1.0')
     def update_policy(self, policy_id: str, policy_type: str, data: Dict):
         request = f'/policy/{policy_type}/{policy_id}'
         url = self._url('config', request)
         return self._put(url, data)
 
-    @MinimumVersionRequired('6.1.0')
+    @minimum_version_required('6.1.0')
     def delete_policy(self, policy_id: str, policy_type: str):
         request = f'/policy/{policy_type}/{policy_id}'
         url = self._url('config', request)
         return self._delete(url)
 
-    @MinimumVersionRequired('6.2.1')
+    @minimum_version_required('6.2.1')
     def create_acp_rule(self, policy_id: str, data: Dict, section=str(), category=str(),
                         insert_before=int(), insert_after=int()):
         request = f'/policy/accesspolicies/{policy_id}/accessrules'
@@ -1127,7 +985,7 @@ class Client(object):
         }
         return self._post(url, data, params)
 
-    @MinimumVersionRequired('6.2.1')
+    @minimum_version_required('6.2.1')
     def create_acp_rules(self, policy_id: str, data: Dict, section=str(), category=str(),
                          insert_before=int(), insert_after=int()):
         request = f'/policy/accesspolicies/{policy_id}/accessrules'
@@ -1140,13 +998,13 @@ class Client(object):
         }
         return self._post(url, data, params)
 
-    @MinimumVersionRequired('6.1.0')
+    @minimum_version_required('6.1.0')
     def get_acp_rule(self, policy_id: str, rule_id: str):
         request = f'/policy/accesspolicies/{policy_id}/accessrules/{rule_id}'
         url = self._url('config', request)
         return self._get(url)
 
-    @MinimumVersionRequired('6.1.0')
+    @minimum_version_required('6.1.0')
     def get_acp_rules(self, policy_id: str):
         request = f'/policy/accesspolicies/{policy_id}/accessrules'
         url = self._url('config', request)
@@ -1155,31 +1013,31 @@ class Client(object):
         }
         return self._get(url, params)
 
-    @MinimumVersionRequired('6.1.0')
+    @minimum_version_required('6.1.0')
     def update_acp_rule(self, policy_id: str, rule_id: str, data: Dict):
         request = f'/policy/accesspolicies/{policy_id}/accessrules/{rule_id}'
         url = self._url('config', request)
         return self._put(url, data)
 
-    @MinimumVersionRequired('6.1.0')
+    @minimum_version_required('6.1.0')
     def delete_acp_rule(self, policy_id: str, rule_id: str):
         request = f'/policy/accesspolicies/{policy_id}/accessrules/{rule_id}'
         url = self._url('config', request)
         return self._delete(url)
 
-    @MinimumVersionRequired('6.2.3')
+    @minimum_version_required('6.2.3')
     def create_autonat_rule(self, policy_id: str, data: Dict):
         request = f'/policy/ftdnatpolicies/{policy_id}/autonatrules'
         url = self._url('config', request)
         return self._post(url, data)
 
-    @MinimumVersionRequired('6.2.3')
+    @minimum_version_required('6.2.3')
     def get_autonat_rule(self, policy_id: str, rule_id: str):
         request = f'/policy/ftdnatpolicies/{policy_id}/autonatrules/{rule_id}'
         url = self._url('config', request)
         return self._get(url)
 
-    @MinimumVersionRequired('6.2.3')
+    @minimum_version_required('6.2.3')
     def get_autonat_rules(self, policy_id: str):
         request = f'/policy/ftdnatpolicies/{policy_id}/autonatrules'
         url = self._url('config', request)
@@ -1188,31 +1046,31 @@ class Client(object):
         }
         return self._get(url, params)
 
-    @MinimumVersionRequired('6.2.3')
+    @minimum_version_required('6.2.3')
     def update_autonat_rule(self, policy_id: str, data: Dict):
         request = f'/policy/ftdnatpolicies/{policy_id}/autonatrules'
         url = self._url('config', request)
         return self._put(url, data)
 
-    @MinimumVersionRequired('6.2.3')
+    @minimum_version_required('6.2.3')
     def delete_autonat_rule(self, policy_id: str, rule_id: str):
         request = f'/policy/ftdnatpolicies/{policy_id}/autonatrules/{rule_id}'
         url = self._url('config', request)
         return self._delete(url)
 
-    @MinimumVersionRequired('6.2.3')
+    @minimum_version_required('6.2.3')
     def create_manualnat_rule(self, policy_id: str, data: Dict):
         request = f'/policy/ftdnatpolicies/{policy_id}/manualnatrules'
         url = self._url('config', request)
         return self._post(url, data)
 
-    @MinimumVersionRequired('6.2.3')
+    @minimum_version_required('6.2.3')
     def get_manualnat_rule(self, policy_id: str, rule_id: str):
         request = f'/policy/ftdnatpolicies/{policy_id}/manualnatrules/{rule_id}'
         url = self._url('config', request)
         return self._get(url)
 
-    @MinimumVersionRequired('6.2.3')
+    @minimum_version_required('6.2.3')
     def get_manualnat_rules(self, policy_id: str):
         request = f'/policy/ftdnatpolicies/manualnatrules/{policy_id}'
         url = self._url('config', request)
@@ -1221,25 +1079,25 @@ class Client(object):
         }
         return self._get(url, params)
 
-    @MinimumVersionRequired('6.2.3')
+    @minimum_version_required('6.2.3')
     def update_manualnat_rule(self, policy_id: str, data: Dict):
         request = f'/policy/ftdnatpolicies/{policy_id}/manualnatrules'
         url = self._url('config', request)
         return self._put(url, data)
 
-    @MinimumVersionRequired('6.2.3')
+    @minimum_version_required('6.2.3')
     def delete_manualnat_rule(self, policy_id: str, rule_id: str):
         request = f'/policy/ftdnatpolicies/{policy_id}/manualnatrules/{rule_id}'
         url = self._url('config', request)
         return self._delete(url)
 
-    @MinimumVersionRequired('6.1.0')
+    @minimum_version_required('6.1.0')
     def create_policy_assignment(self, data: Dict):
         request = '/assignment/policyassignments'
         url = self._url('config', request)
         return self._post(url, data)
 
-    @MinimumVersionRequired('6.1.0')
+    @minimum_version_required('6.1.0')
     def get_policy_assignments(self):
         request = '/assignment/policyassignments'
         url = self._url('config', request)
@@ -1248,13 +1106,13 @@ class Client(object):
         }
         return self._get(url, params)
 
-    @MinimumVersionRequired('6.1.0')
+    @minimum_version_required('6.1.0')
     def get_policy_assignment(self, policy_id: str):
         request = f'/assignment/policyassignments/{policy_id}'
         url = self._url('config', request)
         return self._get(url)
 
-    @MinimumVersionRequired('6.1.0')
+    @minimum_version_required('6.1.0')
     def update_policy_assignment(self, policy_id: str, data: Dict):
         request = f'/assignment/policyassignments/{policy_id}'
         url = self._url('config', request)
