+++ test/auth/test_auth.py
@@ -1,7 +1,9 @@
 import pytest
 
+# -*- coding: utf-8 -*-
+
 from fireREST import Client
-from fireREST import GenericApiError, AuthError, AuthRefreshError
+from fireREST.exceptions import GenericApiError, AuthError, AuthRefreshError
 from requests.auth import HTTPBasicAuth
