+++ test/helper/test_helper.py
@@ -1,3 +1,5 @@
+# -*- coding: utf-8 -*-
+
 import pytest
 
 from fireREST import Client
